myproject/requirements.txt
streamlit
PyPDF2
langchain
langchain_google_genai
google-generativeai
python-dotenv
